chore: lift buildCustomTx to EvmBaseAdapter
