+++ packages/chain-adapters/src/evm/ethereum/index.ts
@@ -1,2 +1,1 @@
 export * from './EthereumChainAdapter'
-export * from './types'
