+++ packages/chain-adapters/src/evm/types.ts
@@ -1,3 +1,5 @@
+import { HDWallet } from '@shapeshiftoss/hdwallet-core'
+import { BIP44Params } from '@shapeshiftoss/types'
 import * as unchained from '@shapeshiftoss/unchained-client'
 
 import { AssetBalance } from '../types'
@@ -36,5 +38,14 @@ export type GetFeeDataInput = {
   from: string
   contractData?: string
 }
+
+export type BuildCustomTxInput = {
+  wallet: HDWallet
+  bip44Params: BIP44Params
+  to: string
+  data: string
+  value: string
+  gasLimit: string
+} & Fees
 
 export type TransactionMetadata = unchained.evm.TxMetadata
